Add support for array as the root object of JSON
Extends the Alamofire extension so that a JSON that has an array as a root
object is considered valid and parses correctly. Additionally, also correctly
parses a single object if the response JSON contains only one dictionary
object in the root array object.

The new additions are unit tested and confirmed working. Some project warnings
were also fixed during the process.